requirements.txt
absl-py==2.1.0
accelerate==0.32.1
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
beautifulsoup4==4.12.3
Brotli==1.1.0
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
diffusers==0.29.2
dlib==19.24.4
easydict==1.13
einops==0.8.0
eval_type_backport==0.2.0
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.2.0
future==1.0.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.64.1
gTTS==2.5.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
insightface==0.7.3
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.4.0
mpmath==1.3.0
mtcnn==0.1.1
mutagen==1.47.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
ollama==0.2.1
onnx==1.16.1
onnxruntime==1.18.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.2.0
pip==24.1.2
prettytable==3.10.2
protobuf==4.25.3
psutil==6.0.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
tbb==2021.11.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.9.0
urllib3==2.2.2
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yt-dlp==2024.7.9
zipp==3.19.2